+++ app/models/map.rb
@@ -14,29 +14,33 @@ class Map
 
   def self.update_existing_users!
     User.all.order(:id).each do |u|
-      if u.state_code.present? or u.country_code.present?
-        Rails.logger.debug "USER ##{u.id} EXISTS: '#{u.state_code}' '#{u.country_code}'"
-        next
-      end
-      results = Geocoder.search(u.current_sign_in_ip)
-      result = results.first
-      if result
-        country = result.country.to_s
-        country = "United States of America" if country == "United States"
-        state_pair = MAP_STATES_AND_CODES.select{|name, code| name.downcase == result.state.to_s.downcase}.first
-        country_pair = MAP_COUNTRIES_AND_CODES.select{|name, code| name.downcase == country.downcase}.first
-        if state_pair
-          u.update(state_code: state_pair[1], country_code: 'us')
-          Rails.logger.debug "USER ##{u.id} UPDATED: '#{u.state_code}' '#{u.country_code}' State: #{result.state}  Country: #{country}"
-        elsif country_pair
-          u.update(country_code: country_pair[1])
-          Rails.logger.debug "USER ##{u.id} UPDATED: '#{u.country_code}' State: #{result.state}  Country: #{country}"
-        else
-          Rails.logger.debug "User ##{u.id} did not find a match for\n    State: #{result.state}\n  Country: #{country}"
-        end
+      self.update_user_location(u)
+    end
+  end
+
+  def self.update_user_location(user)
+    if user.state_code.present? or user.country_code.present?
+      Rails.logger.debug "USER ##{user.id} EXISTS: '#{user.state_code}' '#{user.country_code}'"
+      return
+    end
+    results = ::Geocoder.search(user.current_sign_in_ip)
+    result = results.first
+    if result
+      country = result.country.to_s
+      country = "United States of America" if country == "United States"
+      state_pair = MAP_STATES_AND_CODES.select{|name, code| name.downcase == result.state.to_s.downcase}.first
+      country_pair = MAP_COUNTRIES_AND_CODES.select{|name, code| name.downcase == country.downcase}.first
+      if state_pair
+        user.update(state_code: state_pair[1], country_code: 'us')
+        Rails.logger.debug "USER ##{user.id} UPDATED: '#{user.state_code}' '#{user.country_code}' State: #{result.state}  Country: #{country}"
+      elsif country_pair
+        user.update(country_code: country_pair[1])
+        Rails.logger.debug "USER ##{user.id} UPDATED: '#{user.country_code}' State: #{result.state}  Country: #{country}"
       else
-        Rails.logger.debug "User ##{u.id} no result found for #{u.current_sign_in_ip}"
+        Rails.logger.debug "User ##{user.id} did not find a match for\n    State: #{result.state}\n  Country: #{country}"
       end
+    else
+      Rails.logger.debug "User ##{user.id} no result found for #{user.current_sign_in_ip}"
     end
   end