+++ app/models/surveys/report.rb
@@ -91,7 +91,7 @@ class Report < ActiveRecord::Base
   # - specific user
   #
 
-  def self.frequency_data(question_slug, values=nil)
+  def self.frequency_data(question_slug, values=nil, users=nil)
 
     q = Question.where(slug: question_slug).includes(answer_templates: :answer_options).first
     answer_options = q.answer_templates.first.answer_options
@@ -99,6 +99,7 @@ class Report < ActiveRecord::Base
 
     base_query = Report.where(question_slug: question_slug)
     base_query = base_query.where(value: values.map(&:to_s)) if values.present?
+    base_query = base_query.where(user: users.map(&:to_s)) if users.present?
 
     total_count = base_query.count